requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.4
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
fsspec==2026.2.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
lap==0.5.13
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
polars==1.39.3
polars-runtime-32==1.39.3
psutil==7.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.1
shapely==2.1.2
six==1.17.0
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typing_extensions==4.15.0
ultralytics==8.4.24
ultralytics-thop==2.0.18
urllib3==2.6.3
vidgear==0.3.4
Werkzeug==3.1.6
yt-dlp==2026.3.17
